refacto(app): rename mobile folder to steps
